backend/requirements.txt
# Core Framework
Flask
Flask-Cors
python-dotenv

# AI & Agent Libs
smolagents # Replace with actual package name if different
litellm
openai # Often a dependency for LiteLLM or agents

# ML & Data
h2o
pandas # Use reasonably recent pandas
numpy
ucimlrepo # For fetching the dataset
scikit-learn # Often a dependency for ML tasks or metrics

# Document Processing
PyPDF2
langchain-community # For BM25Retriever, Document, TextSplitter

# Web Search
duckduckgo-search

# Visualization (even if plots generated by agent code)
plotly # Use reasonably recent plotly
matplotlib
kaleido # Often needed by plotly to save static images like PNG

# Ensure versions match your environment if needed, e.g.:
# Flask==2.3.2
# pandas==2.1.0